refactor: Create a new hook for handle the approval
