+++ src/views/Home/components/LotteryCard.tsx
@@ -10,7 +10,7 @@ import { useMultiClaimLottery } from 'hooks/useBuyLottery'
 import { useTotalClaim } from 'hooks/useTickets'
 import BuyModal from 'views/Lottery/components/TicketCard/BuyTicketModal'
 import { useLotteryAllowance } from 'hooks/useAllowance'
-import { useLotteryApprove } from 'hooks/useApprove'
+import { useApproval } from 'hooks/useApproval'
 import PurchaseWarningModal from 'views/Lottery/components/TicketCard/PurchaseWarningModal'
 import CakeWinnings from './CakeWinnings'
 import LotteryJackpot from './LotteryJackpot'
@@ -48,26 +48,11 @@ const FarmedStakingCard = () => {
   const [requesteClaim, setRequestedClaim] = useState(false)
   const TranslateString = useI18n()
   const allowance = useLotteryAllowance()
-  const { onApprove } = useLotteryApprove()
   const [onPresentApprove] = useModal(<PurchaseWarningModal />)
-  const [requestedApproval, setRequestedApproval] = useState(false)
   const { claimAmount } = useTotalClaim()
   const { onMultiClaim } = useMultiClaimLottery()
   const cakeBalance = useTokenBalance(getCakeAddress())
-
-  const handleApprove = useCallback(async () => {
-    try {
-      setRequestedApproval(true)
-      const txHash = await onApprove()
-      // user rejected tx or didn't go thru
-      if (!txHash) {
-        setRequestedApproval(false)
-      }
-      onPresentApprove()
-    } catch (e) {
-      console.error(e)
-    }
-  }, [onApprove, onPresentApprove])
+  const { handleApprove, requestedApproval } = useApproval(onPresentApprove)
 
   const handleClaim = useCallback(async () => {
     try {
